requirements.txt
discord.py>=2.0.0
feedparser
requests
beautifulsoup4
rapidfuzz
databases[postgresql]
asyncpg

